+++ src/main/java/com/besson/endfield/block/custom/PortableOriginiumRigBlock.java
@@ -1,10 +1,20 @@
 package com.besson.endfield.block.custom;
 
 import com.besson.endfield.block.ModBlockEntityWithFacing;
+import com.besson.endfield.blockentity.ModBlockEntities;
 import com.besson.endfield.blockentity.custom.PortableOriginiumRigBlockEntity;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityTicker;
+import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.screen.NamedScreenHandlerFactory;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.Hand;
+import net.minecraft.util.ItemScatterer;
+import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 public class PortableOriginiumRigBlock extends ModBlockEntityWithFacing {
@@ -15,5 +25,36 @@ public class PortableOriginiumRigBlock extends ModBlockEntityWithFacing {
     @Override
     public @Nullable BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
         return new PortableOriginiumRigBlockEntity(pos, state);
+    }
+
+    @Override
+    public @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(World world, BlockState state, BlockEntityType<T> type) {
+        return checkType(type, ModBlockEntities.PORTABLE_ORIGINIUM_RIG,
+                (world1, pos, state1, blockEntity) ->
+                    blockEntity.tick(world1, pos, state1, (PortableOriginiumRigBlockEntity) blockEntity));
+    }
+
+    @Override
+    public void onStateReplaced(BlockState state, World world, BlockPos pos, BlockState newState, boolean moved) {
+        if (state.getBlock() != newState.getBlock()) {
+            BlockEntity blockEntity = world.getBlockEntity(pos);
+            if (blockEntity instanceof PortableOriginiumRigBlockEntity) {
+                ItemScatterer.spawn(world, pos, (PortableOriginiumRigBlockEntity)blockEntity);
+                world.updateComparators(pos, this);
+            }
+            super.onStateReplaced(state, world, pos, newState, moved);
+        }
+    }
+
+    @Override
+    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+        if (!world.isClient()) {
+            NamedScreenHandlerFactory screenHandlerFactory = ((PortableOriginiumRigBlockEntity) world.getBlockEntity(pos));
+            if (screenHandlerFactory != null) {
+                player.openHandledScreen(screenHandlerFactory);
+                return ActionResult.SUCCESS;
+            }
+        }
+        return ActionResult.CONSUME;
     }
 }