+++ src/main/java/com/besson/endfield/blockentity/custom/PortableOriginiumRigBlockEntity.java
@@ -1,10 +1,32 @@
 package com.besson.endfield.blockentity.custom;
 
+import com.besson.endfield.blockentity.ImplementedInventory;
 import com.besson.endfield.blockentity.ModBlockEntities;
+import com.besson.endfield.recipe.custom.PortableOriginiumRigRecipe;
+import com.besson.endfield.screen.custom.PortableOriginiumRigScreenHandler;
+import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerFactory;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.Inventories;
+import net.minecraft.inventory.SimpleInventory;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NbtCompound;
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.network.listener.ClientPlayPacketListener;
+import net.minecraft.network.packet.Packet;
+import net.minecraft.network.packet.s2c.play.BlockEntityUpdateS2CPacket;
+import net.minecraft.screen.PropertyDelegate;
+import net.minecraft.screen.ScreenHandler;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.text.Text;
+import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
 import software.bernie.geckolib.animatable.GeoBlockEntity;
 import software.bernie.geckolib.core.animatable.GeoAnimatable;
 import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
@@ -13,22 +35,197 @@ import software.bernie.geckolib.core.animation.AnimationController;
 import software.bernie.geckolib.core.animation.RawAnimation;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
-public class PortableOriginiumRigBlockEntity extends BlockEntity implements GeoBlockEntity {
+import java.util.Optional;
+
+public class PortableOriginiumRigBlockEntity extends BlockEntity implements GeoBlockEntity, ExtendedScreenHandlerFactory, ImplementedInventory {
 
     private final AnimatableInstanceCache cache = GeckoLibUtil.createInstanceCache(this);
+    private boolean isWorking;
+
+    private final DefaultedList<ItemStack> inventory = DefaultedList.ofSize(1, ItemStack.EMPTY);
+    private static final int OUTPUT_SLOT = 0;
+
+    protected final PropertyDelegate propertyDelegate;
+    private int progress = 0;
+    private int maxProgress = 40;
 
     public PortableOriginiumRigBlockEntity(BlockPos pos, BlockState state) {
         super(ModBlockEntities.PORTABLE_ORIGINIUM_RIG, pos, state);
+        this.propertyDelegate = new PropertyDelegate() {
+            @Override
+            public int get(int index) {
+                return switch (index) {
+                    case 0 -> PortableOriginiumRigBlockEntity.this.progress;
+                    case 1 -> PortableOriginiumRigBlockEntity.this.maxProgress;
+                    default -> 0;
+                };
+            }
+
+            @Override
+            public void set(int index, int value) {
+                switch (index) {
+                    case 0 -> PortableOriginiumRigBlockEntity.this.progress = value;
+                    case 1 -> PortableOriginiumRigBlockEntity.this.maxProgress = value;
+                }
+            }
+
+            @Override
+            public int size() {
+                return 2;
+            }
+        };
+    }
+
+    @Override
+    protected void writeNbt(NbtCompound nbt) {
+        super.writeNbt(nbt);
+        Inventories.writeNbt(nbt, inventory);
+        nbt.putInt("progress", progress);
+        nbt.putBoolean("isWorking", isWorking);
+    }
+
+    @Override
+    public void readNbt(NbtCompound nbt) {
+        super.readNbt(nbt);
+        Inventories.readNbt(nbt, inventory);
+        this.progress = nbt.getInt("progress");
+        this.isWorking = nbt.getBoolean("isWorking");
+    }
+
+    @Override
+    public @Nullable Packet<ClientPlayPacketListener> toUpdatePacket() {
+        return BlockEntityUpdateS2CPacket.create(this);
+    }
+
+    @Override
+    public NbtCompound toInitialChunkDataNbt() {
+        return this.createNbt();
     }
 
     @Override
     public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
         controllers.add(new AnimationController<>(this, "controller", 0,
-                state -> state.setAndContinue(RawAnimation.begin().thenLoop("working"))));
+                state -> this.isWorking
+                        ? state.setAndContinue(RawAnimation.begin().thenLoop("working"))
+                        : state.setAndContinue(RawAnimation.begin().thenLoop("idle"))));
     }
 
     @Override
     public AnimatableInstanceCache getAnimatableInstanceCache() {
         return cache;
+    }
+
+    @Override
+    public DefaultedList<ItemStack> getItems() {
+        return inventory;
+    }
+
+    @Override
+    public void writeScreenOpeningData(ServerPlayerEntity serverPlayerEntity, PacketByteBuf packetByteBuf) {
+        packetByteBuf.writeBlockPos(this.pos);
+    }
+
+    @Override
+    public Text getDisplayName() {
+        return Text.translatable("blockEntity.portable_originium_rig");
+    }
+
+    public boolean isWorking() {
+        return isWorking;
+    }
+
+    public void tick(World world, BlockPos pos, BlockState state, PortableOriginiumRigBlockEntity entity) {
+        if (world.isClient()) {
+            return;
+        }
+
+        boolean activeNow = this.hasCorrectRecipe(world);
+
+        if (this.isOutputSlotAvailable()) {
+            if (activeNow) {
+
+                this.incrementProgress();
+                markDirty(world, pos, state);
+
+                if (this.hasCraftFinished()) {
+                    this.craftItem(world);
+                    this.resetProgress();
+                }
+            } else {
+                this.resetProgress();
+            }
+        } else {
+            this.resetProgress();
+            markDirty(world, pos, state);
+        }
+
+        if (entity.isWorking != activeNow) {
+            entity.isWorking = activeNow;
+            entity.markDirty();
+            world.updateListeners(pos, state, state, 3);
+        }
+    }
+
+    private void resetProgress() {
+        this.progress = 0;
+    }
+
+    private void craftItem(World world) {
+
+        Optional<PortableOriginiumRigRecipe> match = getMatchRecipe(world);
+
+        if (match.isPresent()) {
+            ItemStack result = match.get().getOutput(world.getRegistryManager());
+            this.setStack(OUTPUT_SLOT,
+                    new ItemStack(result.getItem(), getStack(OUTPUT_SLOT).getCount() + result.getCount()));
+        }
+
+    }
+
+    private Optional<PortableOriginiumRigRecipe> getMatchRecipe(World world) {
+        SimpleInventory inv = new SimpleInventory(1);
+        BlockState belowState = world.getBlockState(this.pos.down());
+        ItemStack belowStack = belowState.getBlock().asItem().getDefaultStack();
+        inv.setStack(0, belowStack);
+
+        return world.getRecipeManager()
+                .getFirstMatch(PortableOriginiumRigRecipe.Type.INSTANCE, inv, world)
+                .map(recipe -> (PortableOriginiumRigRecipe) recipe);
+    }
+
+    private boolean hasCraftFinished () {
+        return this.progress >= this.maxProgress;
+    }
+
+    private void incrementProgress() {
+        this.progress++;
+    }
+
+    private boolean hasCorrectRecipe(World world) {
+
+        Optional<PortableOriginiumRigRecipe> match = getMatchRecipe(world);
+
+        if (match.isPresent()) {
+            ItemStack result = match.get().getOutput(world.getRegistryManager());
+            return canInsertAmountIntoOutputSlot(result) && canInsertItemIntoOutputSlot(result.getItem());
+        }
+        return false;
+    }
+
+    private boolean canInsertAmountIntoOutputSlot(ItemStack result) {
+        return this.getStack(OUTPUT_SLOT).getCount() + result.getCount() <= getStack(OUTPUT_SLOT).getMaxCount();
+    }
+
+    private boolean canInsertItemIntoOutputSlot(Item item) {
+        return this.getStack(OUTPUT_SLOT).getItem() == item || this.getStack(OUTPUT_SLOT).isEmpty();
+    }
+
+    private boolean isOutputSlotAvailable() {
+        return this.getStack(OUTPUT_SLOT).isEmpty() || this.getStack(OUTPUT_SLOT).getCount() < this.getStack(OUTPUT_SLOT).getMaxCount();
+    }
+
+    @Override
+    public @Nullable ScreenHandler createMenu(int syncId, PlayerInventory playerInventory, PlayerEntity player) {
+        return new PortableOriginiumRigScreenHandler(syncId, playerInventory, this, this.propertyDelegate);
     }
 }