+++ src/main/java/com/besson/endfield/datagen/ModBlockTagsProvider.java
@@ -1,8 +1,10 @@
 package com.besson.endfield.datagen;
 
+import com.besson.endfield.block.ModBlocks;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
 import net.minecraft.registry.RegistryWrapper;
+import net.minecraft.registry.tag.BlockTags;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -13,6 +15,28 @@ public class ModBlockTagsProvider extends FabricTagProvider.BlockTagProvider {
 
     @Override
     protected void configure(RegistryWrapper.WrapperLookup wrapperLookup) {
+        getOrCreateTagBuilder(BlockTags.PICKAXE_MINEABLE)
+                .add(ModBlocks.AMETHYST_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.COAL_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.COPPER_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.DIAMOND_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.EMERALD_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.GOLD_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.IRON_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.LAPIS_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.ORIGINIUM_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.REDSTONE_MINERAL_VEIN_BLOCK);
 
+        getOrCreateTagBuilder(BlockTags.NEEDS_DIAMOND_TOOL)
+                .add(ModBlocks.AMETHYST_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.COAL_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.COPPER_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.DIAMOND_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.EMERALD_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.GOLD_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.IRON_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.LAPIS_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.ORIGINIUM_MINERAL_VEIN_BLOCK)
+                .add(ModBlocks.REDSTONE_MINERAL_VEIN_BLOCK);
     }
 }