+++ src/main/java/com/besson/endfield/recipe/custom/PortableOriginiumRigRecipe.java
@@ -1,0 +1,90 @@
+package com.besson.endfield.recipe.custom;
+
+import com.google.gson.JsonObject;
+import net.minecraft.inventory.SimpleInventory;
+import net.minecraft.item.ItemStack;
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.recipe.*;
+import net.minecraft.registry.DynamicRegistryManager;
+import net.minecraft.util.Identifier;
+import net.minecraft.util.JsonHelper;
+import net.minecraft.world.World;
+
+public class PortableOriginiumRigRecipe implements Recipe<SimpleInventory> {
+    private final Identifier id;
+    private final Ingredient input;
+    private final ItemStack output;
+
+    public PortableOriginiumRigRecipe(Identifier id, Ingredient input, ItemStack output) {
+        this.id = id;
+        this.input = input;
+        this.output = output;
+    }
+
+    @Override
+    public boolean matches(SimpleInventory inventory, World world) {
+        if (world.isClient()) return false;
+        return input.test(inventory.getStack(0));
+    }
+
+    @Override
+    public ItemStack craft(SimpleInventory inventory, DynamicRegistryManager registryManager) {
+        return output.copy();
+    }
+
+    @Override
+    public boolean fits(int width, int height) {
+        return true;
+    }
+
+    @Override
+    public ItemStack getOutput(DynamicRegistryManager registryManager) {
+        return output;
+    }
+
+    @Override
+    public Identifier getId() {
+        return id;
+    }
+
+    @Override
+    public RecipeSerializer<?> getSerializer() {
+        return Serializer.INSTANCE;
+    }
+
+    @Override
+    public RecipeType<?> getType() {
+        return Type.INSTANCE;
+    }
+
+    public static class Type implements RecipeType<PortableOriginiumRigRecipe>{
+        public static final Type INSTANCE = new Type();
+        public static final String ID = "portable_originium_rig";
+    }
+
+    public static class Serializer implements RecipeSerializer<PortableOriginiumRigRecipe> {
+
+        public static final Serializer INSTANCE = new Serializer();
+        public static final String ID = "portable_originium_rig";
+
+        @Override
+        public PortableOriginiumRigRecipe read(Identifier id, JsonObject json) {
+            Ingredient input = Ingredient.fromJson(json.get("input"));
+            ItemStack output = ShapedRecipe.outputFromJson(JsonHelper.getObject(json, "output"));
+            return new PortableOriginiumRigRecipe(id, input, output);
+        }
+
+        @Override
+        public PortableOriginiumRigRecipe read(Identifier id, PacketByteBuf buf) {
+            Ingredient input = Ingredient.fromPacket(buf);
+            ItemStack output = buf.readItemStack();
+            return new PortableOriginiumRigRecipe(id, input, output);
+        }
+
+        @Override
+        public void write(PacketByteBuf buf, PortableOriginiumRigRecipe recipe) {
+            recipe.input.write(buf);
+            buf.writeItemStack(recipe.output);
+        }
+    }
+}