+++ src/main/java/com/besson/endfield/screen/ModScreens.java
@@ -1,0 +1,19 @@
+package com.besson.endfield.screen;
+
+import com.besson.endfield.ArknightsEndfield;
+import com.besson.endfield.screen.custom.PortableOriginiumRigScreenHandler;
+import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerType;
+import net.minecraft.registry.Registries;
+import net.minecraft.registry.Registry;
+import net.minecraft.screen.ScreenHandlerType;
+import net.minecraft.util.Identifier;
+
+public class ModScreens {
+    public static final ScreenHandlerType<PortableOriginiumRigScreenHandler> PORTABLE_ORIGINIUM_RIG_SCREEN =
+            Registry.register(Registries.SCREEN_HANDLER, new Identifier(ArknightsEndfield.MOD_ID, "portable_originium_rig_screen"),
+                    new ExtendedScreenHandlerType<>(PortableOriginiumRigScreenHandler::new));
+
+    public static void register() {
+
+    }
+}